requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
pillow==10.0.0
torch==2.0.1
torchvision==0.15.2
tensorflow==2.13.0
numpy==1.24.3
python-dotenv==1.0.0
opencv-python-headless==4.7.0.72
python-magic==0.4.27
aiofiles==23.1.0
jinja2==3.1.2
